requirements.txt
feedparser==6.0.11
httpx==0.27.0
pydantic-settings==2.3.0
python-dotenv==1.0.1
tenacity==8.3.0
notion-client==2.2.1
slack-sdk==3.30.0
openai==1.55.3
apscheduler==3.10.4
pyyaml==6.0.1
pytest==8.2.0
